NewsRAG/pyproject.toml
[tool.poetry]
name = "newsrag"
version = "0.1.0"
description = ""
authors = ["MattRossi <mattrossi.phy@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.11,<3.13"
torch = {version = "2.0.1+cu118", source = "torch"}
ipykernel = "^6.29.4"
pymongo = "^4.7.3"
langchain = "0.2.0"
python-dotenv = "^1.0.1"
worldnewsapi = "^1.1.1"
pandas = "^2.2.2"
sentence-transformers = "^3.0.0"
pinecone-text = "^0.9.0"
langchain-pinecone = "^0.1.1"
langchain-openai = "^0.1.8"
langchain-community = "^0.2.4"
langchain-huggingface = "^0.0.3"
logging = "^0.4.9.6"
langchain-mongodb = "^0.1.6"
streamlit = "^1.35.0"
spacy = "^3.7.5"
keybert = "^0.8.4"
matplotlib = "^3.9.0"
seaborn = "^0.13.2"
wordcloud = "^1.9.3"

[[tool.poetry.source]]
name = "PyPI"
priority = "primary"


[[tool.poetry.source]]
name = "torch"
url = "https://download.pytorch.org/whl/cu118"
priority = "supplemental"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
